+++ .pdkignore
@@ -22,10 +22,3 @@
 /convert_report.txt
 /update_report.txt
 .DS_Store
-bin
-import
-.tmp
-.rbenv-gemsets
-.ruby-version
-*.lock
-!build/vendor/
