+++ .rubocop_todo.yml
@@ -1,0 +1,12 @@
+# This configuration was generated by
+# `rubocop --auto-gen-config`
+# on 2018-08-13 10:44:38 +0800 using RuboCop version 0.49.1.
+# The point is for the user to remove these configuration records
+# one by one as the offenses are removed from the code base.
+# Note that changes in the inspected code, or installation of new
+# versions of RuboCop, may require this file to be generated again.
+
+# Offense count: 3
+RSpec/LetBeforeExamples:
+  Exclude:
+    - 'spec/integration/puppet_x/puppetlabs/powershell_manager_spec.rb'
