+++ metadata.json
@@ -7,10 +7,11 @@
   "source": "https://github.com/puppetlabs/puppetlabs-dsc",
   "project_page": "https://github.com/puppetlabs/puppetlabs-dsc",
   "issues_url": "https://tickets.puppet.com/browse/MODULES",
-  "tags": [
-    "powershell",
-    "desired-state-configuration",
-    "dsc"
+  "dependencies": [
+    {
+      "name": "puppetlabs/reboot",
+      "version_requirement": ">= 1.2.1 < 3.0.0"
+    }
   ],
   "operatingsystem_support": [
     {
@@ -33,10 +34,12 @@
       "version_requirement": ">= 4.7.0 < 6.0.0"
     }
   ],
-  "dependencies": [
-    {
-      "name": "puppetlabs/reboot",
-      "version_requirement": ">= 1.2.1 < 3.0.0"
-    }
-  ]
+  "tags": [
+    "powershell",
+    "desired-state-configuration",
+    "dsc"
+  ],
+  "pdk-version": "1.6.0",
+  "template-url": "file://C:/Program Files/Puppet Labs/DevelopmentKit/share/cache/pdk-templates.git",
+  "template-ref": "1.6.0-0-gf5564c0"
 }
