+++ spec/spec_helper.rb
@@ -1,66 +1,45 @@
+
 require 'puppetlabs_spec_helper/module_spec_helper'
-require 'rake'
-require 'fileutils'
+require 'rspec-puppet-facts'
 
-# for debugging
-require 'pry' if Bundler.rubygems.find_name('pry').any?
+begin
+  require 'spec_helper_local' if File.file?(File.join(File.dirname(__FILE__), 'spec_helper_local.rb'))
+rescue LoadError => loaderror
+  warn "Could not require spec_helper_local: #{loaderror.message}"
+end
 
-fixture_path = File.expand_path(File.join(__FILE__, '..', 'fixtures'))
+include RspecPuppetFacts
 
-# Workaround for cross-module type dependency
-dsc_lib_path = File.join(fixture_path, 'modules', 'dsc','lib')
-$:.unshift(dsc_lib_path) unless $:.include?(dsc_lib_path)
+default_facts = {
+  puppetversion: Puppet.version,
+  facterversion: Facter.version,
+}
+
+default_facts_path = File.expand_path(File.join(File.dirname(__FILE__), 'default_facts.yml'))
+default_module_facts_path = File.expand_path(File.join(File.dirname(__FILE__), 'default_module_facts.yml'))
+
+if File.exist?(default_facts_path) && File.readable?(default_facts_path)
+  default_facts.merge!(YAML.safe_load(File.read(default_facts_path)))
+end
+
+if File.exist?(default_module_facts_path) && File.readable?(default_module_facts_path)
+  default_facts.merge!(YAML.safe_load(File.read(default_module_facts_path)))
+end
 
 RSpec.configure do |c|
-  c.module_path = File.join(fixture_path, 'modules')
-  c.manifest_dir = File.join(fixture_path, 'manifests')
-  c.pattern = FileList[c.pattern].exclude(/^spec\/fixtures/)
-
-  # ensure specs are cleaned up
-  tmpdir = Dir.mktmpdir("rspecrun_dsc")
-  oldtmpdir = Dir.tmpdir()
-  ENV['TMPDIR'] = tmpdir
-
-  c.expect_with :rspec do |e|
-    e.syntax = [:should, :expect]
-  end
-
-  c.after :suite do
-    # return to original tmpdir
-    ENV['TMPDIR'] = oldtmpdir
-    if Puppet::Util::Platform.windows?
-      #FileUtils.rm_rf(tmpdir) - follows symlinks on Windows
-      # and deletes there as well :/
-      system("rmdir /s /q \"#{tmpdir}\"")
-    end
+  c.default_facts = default_facts
+  c.before :each do
+    # set to strictest setting for testing
+    # by default Puppet runs at warning level
+    Puppet.settings[:strict] = :warning
   end
 end
 
-# We need this because the RAL uses 'should' as a method.  This
-# allows us the same behaviour but with a different method name.
-class Object
-  alias :must :should
-  alias :must_not :should_not
-end
-
-Puppet::Util::Log.level = :warning
-Puppet::Util::Log.newdestination(:console)
-
-# Mocking Objects
-class MockCIMClass
-  attr_accessor :name
-  attr_accessor :qualifiers
-
-  def initialize(name)
-    @name = name
-    @qualifiers = {}
+def ensure_module_defined(module_name)
+  module_name.split('::').reduce(Object) do |last_module, next_module|
+    last_module.const_set(next_module, Module.new) unless last_module.const_defined?(next_module)
+    last_module.const_get(next_module)
   end
 end
 
-class MockCIMQualifier
-  attr_accessor :value
-
-  def initialize(value)
-    @value = value
-  end
-end
+# 'spec_overrides' from sync.yml will appear below this line
